+++ Tests/SwiftLintFrameworkTests/LinterCacheTests.swift
@@ -104,26 +104,6 @@ class LinterCacheTests: XCTestCase {
         XCTAssertEqual(cache.violations(forFile: file2, configuration: initialConfig)!, [], file: file, line: line)
     }
 
-    // MARK: Cache Initialization
-
-    func testInitThrowsWhenUsingInvalidCacheFormat() {
-        let cache = [["version": "0.1.0"]]
-        checkError(LinterCacheError.invalidFormat) {
-            _ = try LinterCache(cache: cache)
-        }
-    }
-
-    func testSaveThrowsWithNoLocation() throws {
-        let cache = try LinterCache(cache: [:])
-        checkError(LinterCacheError.noLocation) {
-            try cache.save()
-        }
-    }
-
-    func testInitSucceeds() {
-        XCTAssertNotNil(try? LinterCache(cache: [:]))
-    }
-
     // MARK: Cache Reuse
 
     // Two subsequent lints with no changes reuses cache
@@ -360,19 +340,5 @@ class LinterCacheTests: XCTestCase {
             let version = "4.0.0" // Since we can't pass SWIFT_VERSION=3 to sourcekit, it returns 4.0.0
         #endif
         XCTAssertEqual(SwiftVersion.current.rawValue, version)
-    }
-
-    // MARK: JSON output
-
-    func testCacheToJSONDoesntCrash() {
-        // swiftlint:disable line_length
-        let key1 = "[\"/SwiftLint/source\",[[\"block_based_kvo\",\"warning\"],[\"class_delegate_protocol\",\"warning\"],[\"closing_brace\",\"warning\"],[\"closure_parameter_position\",\"warning\"],[\"colon\",\"warning, flexible_right_spacing: false, apply_to_dictionaries: true\"],[\"comma\",\"warning\"],[\"compiler_protocol_init\",\"warning\"],[\"control_statement\",\"warning\"],[\"custom_rules\",\"\"],[\"cyclomatic_complexity\",\"warning: 10, error: 20, ignores_case_statements: false\"],[\"discarded_notification_center_observer\",\"warning\"],[\"discouraged_direct_init\",\"warning, types: [\"Bundle\", \"Bundle.init\", \"UIDevice\", \"UIDevice.init\"]\"],[\"dynamic_inline\",\"error\"],[\"empty_enum_arguments\",\"warning\"],[\"empty_parameters\",\"warning\"]]]"
-
-        let key2 = "[\"/SwiftLint/Source\",[[\"block_based_kvo\",\"warning\"],[\"class_delegate_protocol\",\"warning\"],[\"closing_brace\",\"warning\"],[\"closure_parameter_position\",\"warning\"],[\"colon\",\"warning, flexible_right_spacing: false, apply_to_dictionaries: true\"],[\"comma\",\"warning\"],[\"compiler_protocol_init\",\"warning\"],[\"control_statement\",\"warning\"],[\"custom_rules\",\"\"],[\"cyclomatic_complexity\",\"warning: 10, error: 20, ignores_case_statements: false\"],[\"discarded_notification_center_observer\",\"warning\"],[\"discouraged_direct_init\",\"warning, types: [\"Bundle\", \"Bundle.init\", \"UIDevice\", \"UIDevice.init\"]\"],[\"dynamic_inline\",\"error\"],[\"empty_enum_arguments\",\"warning\"],[\"empty_parameters\",\"warning\"]]]"
-        // swiftlint:enable line_length
-
-        let dict = [key1: "test", key2: "test2"]
-
-        XCTAssertNoThrow(try dict.toJSON())
     }
 }